pyproject.toml
[project]
name = "mcp-google-ads"
version = "0.1.0"
description = "MCP Server for Google Ads API v23 - Full CRUD by Biomo"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [
    { name = "Thiago Bauer", email = "thiago@biomo.com.br" },
]
dependencies = [
    "google-ads>=28.0.0",
    "mcp[cli]>=1.2.0",
    "pydantic>=2.0.0",
]

[project.scripts]
mcp-google-ads = "mcp_google_ads.server:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-mock>=3.14",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_google_ads"]

[tool.hatch.build.targets.sdist]
include = ["src/mcp_google_ads"]
